+++ R/data_group.R
@@ -6,7 +6,7 @@
 #' grouping information from a grouped data frame.
 #'
 #' @param data A data frame
-#' @inheritParams find_columns
+#' @inheritParams extract_column_names
 #'
 #' @return A grouped data frame, i.e. a data frame with additional information
 #' about the grouping structure saved as attributes.
@@ -41,26 +41,26 @@ data_group <- function(data,
     verbose = verbose
   )
   # create grid with combinations of all levels
-  grid <- as.data.frame(expand.grid(lapply(data[select], unique)))
+  my_grid <- as.data.frame(expand.grid(lapply(data[select], unique)))
   # sort grid
-  grid <- grid[do.call(order, grid), , drop = FALSE]
+  my_grid <- my_grid[do.call(order, my_grid), , drop = FALSE]
 
-  .rows <- lapply(seq_len(nrow(grid)), function(i) {
+  .rows <- lapply(seq_len(nrow(my_grid)), function(i) {
     as.integer(data_match(
       data,
-      to = grid[i, , drop = FALSE],
+      to = my_grid[i, , drop = FALSE],
       match = "and",
       return_indices = TRUE,
       drop_na = FALSE
     ))
   })
-  grid[[".rows"]] <- .rows
+  my_grid[[".rows"]] <- .rows
 
   # remove data_match attributes
-  attr(grid, "out.attrs") <- NULL
-  attr(grid, ".drop") <- TRUE
+  attr(my_grid, "out.attrs") <- NULL
+  attr(my_grid, ".drop") <- TRUE
 
-  attr(data, "groups") <- grid
+  attr(data, "groups") <- my_grid
   class(data) <- unique(c("grouped_df", "data.frame"), class(data))
 
   data