+++ R/data_rescale.R
@@ -4,7 +4,7 @@
 #' (change the keying/scoring direction), or to expand a range.
 #'
 #' @inheritParams categorize
-#' @inheritParams find_columns
+#' @inheritParams extract_column_names
 #' @inheritParams standardize.data.frame
 #'
 #' @param to Numeric vector of length 2 giving the new range that the variable
@@ -202,7 +202,7 @@ rescale.grouped_df <- function(x,
   # create the new variables and updates "select", so new variables are processed
   if (!isFALSE(append)) {
     # process arguments
-    args <- .process_append(
+    my_args <- .process_append(
       x,
       select,
       append,
@@ -210,8 +210,8 @@ rescale.grouped_df <- function(x,
       preserve_value_labels = TRUE
     )
     # update processed arguments
-    x <- args$x
-    select <- args$select
+    x <- my_args$x
+    select <- my_args$select
   }
 
   x <- as.data.frame(x)
@@ -263,15 +263,15 @@ rescale.data.frame <- function(x,
   # create the new variables and updates "select", so new variables are processed
   if (!isFALSE(append)) {
     # process arguments
-    args <- .process_append(
+    my_args <- .process_append(
       x,
       select,
       append,
       append_suffix = "_r"
     )
     # update processed arguments
-    x <- args$x
-    select <- args$select
+    x <- my_args$x
+    select <- my_args$select
   }
 
   # Transform the range so that it is a list now