+++ R/datawizard-package.R
@@ -7,12 +7,12 @@
 #' A lightweight package to assist in key steps involved in any data analysis
 #' workflow:
 #'
-#' - wrangling the raw data to get it in the needed form,
-#' - applying preprocessing steps and statistical transformations, and
-#' - compute statistical summaries of data properties and distributions.
+#'  - wrangling the raw data to get it in the needed form,
+#'  - applying preprocessing steps and statistical transformations, and
+#'  - compute statistical summaries of data properties and distributions.
 #'
 #' It is also the data wrangling backend for packages in 'easystats' ecosystem.
-#' References: Patil et al. (2022) \doi{10.21105/joss.04684}.
+#' Reference: Patil et al. (2022) \doi{10.21105/joss.04684}.
 #'
 #' @docType package
 #' @aliases datawizard datawizard-package