+++ tests/testthat/test-standardize_models.R
@@ -57,7 +57,6 @@ test_that("transformations", {
     ignore_attr = TRUE
   )
 
-  skip_if_not_installed("insight", minimum_version = "0.10.0")
   d <- data.frame(
     time = as.factor(c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5)),
     group = c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2),
@@ -303,6 +302,7 @@ test_that("brms", {
   skip_on_cran()
   skip_on_os(c("windows", "mac"))
   skip_if_not_installed("brms")
+  skip_if_not_installed("RcppEigen")
 
   invisible(
     capture.output({